Refactor ControlMethodArgs
