+++ lib/view_component/storybook.rb
@@ -12,6 +12,7 @@ module ViewComponent
     autoload :StoryConfig
     autoload :ControlMethodArgs
     autoload :Dsl
+    autoload :MethodArgs
 
     include ActiveSupport::Configurable
     # Set the location of component previews through app configuration:
